+++ tests/testthat/test-arrange.r
@@ -109,3 +109,19 @@ test_that("grouped arrange sorts first by group (#491)", {
 
   expect_equal(df1, df2)
 })
+
+test_that("arrange keeps the grouping structure (#605)", {
+  dat <- data_frame(x = 4:1, g = c('b','b','a','a'))
+  res <- dat %>% group_by(g) %>% arrange()
+  expect_is(res, "grouped_df" )
+  expect_false(is.unsorted(res$g))
+  expect_equal(res$x, c(2,1,4,3))
+  expect_equal(res$g, c("a", "a", "b", "b"))
+  
+  res <- dat %>% group_by(g) %>% arrange(x)
+  expect_is(res, "grouped_df")
+  expect_false(is.unsorted(res$g))
+  expect_true(all(summarise(res, sorted = ! is.unsorted(x) )$sorted))
+  expect_equal(attr(res,"indices"), list( c(0,1), c(2,3)) )
+})
+