+++ NEWS.md
@@ -1,11 +1,13 @@
 # dplyr 0.2.0.9000
 
+* `arrange` keeps the grouping structure of grouped data (#605). 
+
 * `left_join` was confused about by variables for x and y (#617). 
 
 * dplyr now depends on Lahman 3.0.1. A number of examples have been updated
   to reflect modified field names (#586).
 
-* `contains()` accidentally matched regualar expression now it passes
+* `contains()` accidentally matched regular expression now it passes
   `fixed = TRUE` to `grep()` (#608).
 
 * `src_monetdb()` is now implemented in MonetDB.R, not dplyr.